conversational-memory-agent/memory/long_term.py
"""
Long-term memory with semantic search capabilities.

TODO: Complete this module to implement semantic vector store.
"""

import logging
import numpy as np
from typing import (
    Any,
    Dict,
    List,
)
from sklearn.feature_extraction.text import TfidfVectorizer
from sklearn.metrics.pairwise import cosine_similarity


logging.basicConfig(
    level=logging.INFO,
    format="%(asctime)s,p%(process)s,{%(filename)s:%(lineno)d},%(levelname)s,%(message)s",
)

logger = logging.getLogger(__name__)


class SemanticMemory:
    """Long-term memory with semantic search using embeddings.

    This class stores conversation summaries with their vector embeddings
    and enables semantic search to retrieve relevant past context.
    """

    def __init__(self):
        """Initialize semantic memory store."""
        # TODO: Initialize storage structures
        # Hint: You'll need:
        # - List to store summaries
        # - List/array to store embeddings
        # - Metadata (turn ranges, timestamps)
        # - TfidfVectorizer for generating embeddings

        # TODO: Initialize vectorizer
        # Hint: TfidfVectorizer(max_features=100) is a good start

        logger.info("Initialized long-term semantic memory")


    def add_summary(
        self,
        summary: str,
        turn_range: tuple[int, int]
    ) -> None:
        """Store a summary with its embedding.

        This method:
        1. Generates an embedding for the summary
        2. Stores the summary, embedding, and metadata

        Args:
            summary: Compressed conversation summary text
            turn_range: (start_turn, end_turn) for this summary

        Example:
            memory.add_summary(
                "User Alice discussed programming interests",
                (1, 10)
            )
        """
        # TODO: Generate embedding for the summary
        # Hint: Use _generate_embedding() method

        # TODO: Store summary with its embedding and metadata
        # Hint: Create a dictionary with:
        # {
        #     "summary": summary text,
        #     "embedding": vector,
        #     "turn_range": (start, end),
        #     "timestamp": current time
        # }

        # TODO: Update the vectorizer if needed
        # Hint: You may need to refit vectorizer with all summaries

        logger.info(f"Added summary for turns {turn_range[0]}-{turn_range[1]}")


    def search(
        self,
        query: str,
        top_k: int = 3
    ) -> List[Dict[str, Any]]:
        """Search for semantically relevant summaries.

        This method:
        1. Generates embedding for the query
        2. Computes similarity with all stored summaries
        3. Returns top-k most similar summaries

        Args:
            query: Search query (typically the current user message)
            top_k: Number of most relevant summaries to return

        Returns:
            List of dictionaries containing relevant summaries and metadata

        Example:
            results = memory.search("What programming languages?", top_k=2)
            # Returns summaries most relevant to programming languages
        """
        # TODO: Handle empty memory case
        # Hint: Return empty list if no summaries stored

        # TODO: Generate embedding for query
        # Hint: Use _generate_embedding() method

        # TODO: Compute cosine similarity between query and all stored embeddings
        # Hint: Use sklearn.metrics.pairwise.cosine_similarity

        # TODO: Get top-k most similar summaries
        # Hint: Use np.argsort to get indices of highest similarities

        # TODO: Return list of relevant summary dictionaries

        logger.info(f"Retrieved {top_k} relevant summaries for query")
        return []


    def _generate_embedding(
        self,
        text: str
    ) -> np.ndarray:
        """Generate embedding vector for text.

        This is a simple implementation using TF-IDF. In production,
        you might use sentence transformers or API-based embeddings.

        Args:
            text: Text to embed

        Returns:
            Numpy array representing the text embedding
        """
        # TODO: Generate TF-IDF embedding
        # Hint: Use self.vectorizer.transform([text]).toarray()[0]

        # TODO: Handle case where vectorizer hasn't been fit yet
        # Hint: Fit on [text] if needed, or return zero vector

        return np.zeros(100)  # Placeholder


    def get_summary_count(self) -> int:
        """Get the number of summaries stored.

        Returns:
            Number of summaries in long-term memory
        """
        # TODO: Implement
        return 0


    def clear(self) -> None:
        """Clear all summaries from long-term memory."""
        # TODO: Reset all storage structures
        logger.info("Cleared long-term memory")
