pyproject.toml
[project]
name = "agents-memory"
version = "0.1.0"
description = "Agent memory patterns lab: From circular buffers to hierarchical memory systems"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.40.0",
    "boto3>=1.34.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "numpy>=1.26.0",
    "scikit-learn>=1.4.0",
    "rich>=13.0.0",
]

[build-system]
requires = [
    "hatchling",
]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/agents_memory"]
